LPS-111943 Rename file to ConnectionContext
