+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/components/TotalCount.js
@@ -13,7 +13,7 @@ import {useIsMounted} from 'frontend-js-react-web';
 import PropTypes from 'prop-types';
 import React, {useContext, useEffect, useState} from 'react';
 
-import ConnectionContext from '../context/connection';
+import ConnectionContext from '../context/ConnectionContext';
 import {useAddWarning} from '../context/store';
 import Hint from './Hint';
